Move extension method to appropriate class
