+++ RoboSharp/Extensions/ISourceDestinationExtensions.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -10,7 +11,7 @@ namespace RoboSharp.Extensions
     /// <summary>
     /// Extension Methods for the <see cref="IDirectorySourceDestinationPair"/> interface
     /// </summary>
-    public static class IDirectorySourceDestinationPairExtensions
+    public static class ISourceDestinationPairExtensions
     {
         private static bool InverseAny<T>(this IEnumerable<T> collection, Func<T, bool> match) => !collection.Any(match);
 
@@ -25,6 +26,95 @@ namespace RoboSharp.Extensions
         public static bool IsLocatedOnSameDrive(this IFileSourceDestinationPair pair) => Path.GetPathRoot(pair.Source.FullName) == Path.GetPathRoot(pair.Destination.FullName);
 
         #endregion
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static bool FileDoesntExist(string destination) => !File.Exists(destination);
+
+        #region < IsSourceNewer >
+
+        /// <summary>
+        /// Check if the Source file is newer than the Destination File
+        /// </summary>
+        /// <param name="source">This path is assumed to exist since it is the patht to copy from</param>
+        /// <param name="destination">the destination path - may or may not exist</param>
+        /// <returns>TRUE if the source is newer, or the destination does not exist. FALSE if the destination is newer.</returns>
+        public static bool IsSourceNewer(string source, string destination)
+        {
+            if (FileDoesntExist(destination)) return true;
+            return File.GetLastWriteTime(source) > File.GetLastWriteTime(destination);
+        }
+        /// <inheritdoc cref="IsSourceNewer(string, string)"/>
+        public static bool IsSourceNewer(FileInfo source, FileInfo destination)
+        {
+            if (!destination.Exists) return source.Exists;
+            return source.LastWriteTime > destination.LastWriteTime;
+        }
+        /// <inheritdoc cref="IsSourceNewer(string, string)"/>
+        public static bool IsSourceNewer(this IFileSourceDestinationPair copier)
+        {
+            return IsSourceNewer(copier.Source, copier.Destination);
+        }
+
+        #endregion
+
+        #region < IsDestinationNewer >
+
+        /// <summary>
+        /// Check if the Destination file is newer than the Source file
+        /// </summary>
+        /// <param name="source">This path is assumed to exist since it is the patht to copy from</param>
+        /// <param name="destination">the destination path - may or may not exist</param>
+        /// <returns>TRUE if the destination file is newer, otherwise false</returns>
+        public static bool IsDestinationNewer(string source, string destination)
+        {
+            if (FileDoesntExist(destination)) return false;
+            return File.GetLastWriteTime(source) < File.GetLastWriteTime(destination);
+        }
+        /// <inheritdoc cref="IsDestinationNewer(string, string)"/>
+        public static bool IsDestinationNewer(FileInfo source, FileInfo destination)
+        {
+            if (!destination.Exists) return false;
+            return source.LastWriteTime < destination.LastWriteTime;
+        }
+        /// <inheritdoc cref="IsDestinationNewer(string, string)"/>
+        public static bool IsDestinationNewer(this IFileSourceDestinationPair copier)
+        {
+            return IsDestinationNewer(copier.Source, copier.Destination);
+        }
+
+        #endregion
+
+        #region < IsSameDate >
+
+        /// <summary>
+        /// Check if the Source and Destination files have the same LastWriteTime
+        /// </summary>
+        /// <param name="source">This path is assumed to exist since it is the patht to copy from</param>
+        /// <param name="destination">the destination path - may or may not exist</param>
+        /// <returns>TRUE if both files exist, and their LastWriteTime is identical</returns>
+        public static bool IsSameDate(string source, string destination)
+        {
+            if (FileDoesntExist(source)) return false;
+            if (FileDoesntExist(destination)) return false;
+            return File.GetLastWriteTime(source) < File.GetLastWriteTime(destination);
+        }
+
+        /// <inheritdoc cref="IsDestinationNewer(string, string)"/>
+        public static bool IsSameDate(FileInfo source, FileInfo destination)
+        {
+            if (!source.Exists) return false;
+            if (!destination.Exists) return false;
+            return source.LastWriteTime == destination.LastWriteTime;
+        }
+
+        /// <inheritdoc cref="IsDestinationNewer(string, string)"/>
+        public static bool IsSameDate(this IFileSourceDestinationPair copier)
+        {
+            return IsDestinationNewer(copier.Source, copier.Destination);
+        }
+
+        #endregion
+
 
         #region < Create Pair Functions >
